requirements.txt
douban-client
sinaweibopy
pymongo
bson
Fabric
PyRSS2Gen
markdown
